Enable / disable Morello in build scripts
A new option --enable-morello / --disable-morello allows to
switch Morello on / off (by default Morello is enabled).

The parts of the makefiles related to Morello are moved
under arc/aarch64 because they are only relevant for this
target.

The configure script will check if Morello can be enabled:
we must have AArch64 target and the toolchain must support
flags related to Morello, otherwise configure step fails
with a clear message.

Change-Id: I2e959720741d4ed3471191fb6fa47bc397aff72f