Port to Morello: disable VDSO
Port clock_gettime, clock_getres, gettimeofday

Until we have a kernel which supports the VDSO, there is no point
trying to use it from Morello code. This patch disables use of VDSO
and ports three of the four AArch64 VDSO functions to Morello.

The fourth function, rt_sigreturn, cannot be tested in isolation, so
it would be better to leave porting rt_sigreturn until we look into
porting signal handlers.

Change-Id: Ib0e1fc050171cccd4db67f2cb1061194a4685800