Morello: Enable malloc and free
This is porting the minimum to get malloc and free to work. malloc-related
functions like realloc are not yet ported. It must also be noted that
malloc return a wide capability, one whose bound allow tamptering with
the metadata and with some adjacent malloc region. Narrowing this bound
require some non negligeable work, because free needs to access the
metadata.

Change-Id: I6633f7a81afcaebcae8e9dc3364cb60708403f89