Morello: Port more pthread internal structures to Morello
Fixes sizing on some internal structures of pthread when compiling for
Morello (cond vars, mutexes, etc). This should now allow the use of
condition variables and mutexes, and one of the libc-tests has been
enabled correspondingly. Temporarily, the largest size for arrays within
the unions is always used, due to difficulty changing the size
conditional on the definition of MORELLO - however, this can be fixed
once all objects are being compiled for Morello within Musl.

Also fixes a couple of bugs with the implementation of __clone on
Morello.

Change-Id: If9011b97ac2f3c9094278b281dd53e9322f9d912