Morello: Make the malloc's returned pointer morello-aligned
This is the first step to narrow this pointer because Morello's bounds
have a finite resolution. Setting some bounds require some alignment
dependent on the length of the bounds.

To align the user's pointer, we use musl's offset mechanism. This had
two purposes: reducing address reuse, and making aligned-malloc possible
By tweaking this mechanism (and oversizing the allocation request) we
can ensure the proper alignment for any size alloc-ed without extensive
changes to musl's malloc.

Change-Id: I2251e662e80534a8cf38b178b51b5e8d9d53a683