Morello: Enable the bounds narrowing in malloc
Building upon the previous commit, this one finally narrows the bounds
of the returned capability. This commit updates a few old tests so that
they check for the correct bounds. The capability is narrowed, but not
yet safe. Some permissions needs to be removed, and some additional
checks need to be introduced (in particular checking that the capability
is pristine and not just part of the original one returned by malloc).

Change-Id: I2fdea9ea2cf728f845829d5fbcbc64b7feb7dc1d