Split UNIT and GRP_SIZE, fix realloc len
UNIT has two meanings in the original code: both the minimum
alignment of allocations, and the size of the group struct.
On CHERI, the group struct becomes larger due to the increased
size of the pointer member. However, the alignment requirement
does not need to increase. This patch splits these two meanings
into different defines, so we can have the larger group size
with the original alignment requirement (16).

Fixes a bug where the decrease in UNIT caused there to not be
enough padding to align larger allocations correctly, due to
incorrect padding calculation in malloc().

Introduces tests for malloc() and realloc() which cover all
sizeclasses for coverage, and adds some extra pointer checks
in other tests.

Also fixes a minor bug with realloc where the allocation length
was slightly longer than it needs to be, due to malloc() (in the
malloc-then-free code path) receiving the a length which was
rounded with cheri_round_representable_length(), instead of the
raw user-requested length.

Change-Id: I461a3edca489c04fff03cc59ff707494bc6fd8e8