Morello: Fix a couple of malloc/realloc bugs
Some of the logic around calculating sizes of allocations was incorrect.

In the case of malloc, UNIT was used where GRP_SIZE should have been
used (2*UNIT on Morello), and allocations could be too small and cause
out-of-bounds faults.
In the case of realloc, when falling back to malloc()+free() in the
case where alignment *did not* change, the size given to malloc() was
already rounded up (which malloc() then does again internally) and the
allocation was slightly larger than necessary (this would not cause any
faults or real issues).

Change-Id: I59ba7b80a77e36042a43affe5ec96b7de181792e