Morello: Fix out of bounds access in mbsrtowcs
Previously, the bitwise operations would detect the null character at
the end of the string by reading the next 4 chars, even if they were
out of bounds. This is not allowed in Morello, so a bounds check is
added before the access. If the next 4 chars access would cross the
limit of the capability, the algorithm resumes to handling each char
independently.

Change-Id: I305b219f9e9e7b9616cb66cb36425b6e4e84f887