[CHERIseed] Replace memcpy family when CHERI is used
This patch replaces the arch-specific assembly implementations of
memcpy, memmove and memset for a capability aware one written in C
when in pure-cap. In hybrid-cap, both implementations are available
and the capability-aware one has a `_c` suffix.

The capability-aware implementations are very bare bones and
not optimized but should be sufficient for now.

Co-authored-by: Tamas Petz <tamas.petz@arm.com>

Change-Id: Id517c00acff1370b00e4ec5d788ab0582b27921d