Morello: Fix mallocng
 - Fix size_classes for Morello
 - Fix size_to_class for Morello

In mallocng, size classes are selected to minimize the remainder in a
page, with every fourth having zero remainder for allocation of a group
with one slot. For Morello, the size of the group header is doubled,
causing the size classes to be misaligned with page boundaries, causing
errors where alignment is assumed.

Size classes are adjusted to realign with page boundaries, and a check
added to prevent out-of-slot offsets.

Conversion of malloc length to size class is fixed to account for the
increased group header size for Morello.

Change-Id: Id22d66d78d3c027f539327fe4b34b980a0ea3da8