Morello: fix alignment of siginfo_t struct members
In Morello, where pointers are stored as 128 bit capabilities, the
siginfo_t->__sigchld struct member offset (0x20) is different to that
produced by linux/include/uapi/asm-generic/siginfo.h (0x18).

This is caused by the musl definition including sigval in a union with
__sigchld. As sigval may contain a ptr, it is therefore aligned on ptr
boundaries. In the case of 16B caps/ptrs, this causes __sigchld to also
be aligned at a 16B boundary. This behaviour is not consistent with the
kernel uapi definition.

Remove problematic unions from musl definition to ensure offsets are
consistent with kernel uapi.

Change-Id: I44ca80383300846d5ff67eb9777087412a00057e