Morello: [epoll] return -EINVAL from epoll_create() if size is non-positive
The man page for epoll_create() states that the 'size' argument must be
positive, otherwise EINVAL is returned. musl currently ignores the
argument and does not return EINVAL. Change it to match the man page.

Worth noting that this is needed for an LTP (Linux Test Project) test to
pass (epoll_create02).

Change-Id: Iba1034d8ca509c67e31ca8f805ef6570fa6d0884