Morello: Pass stdin to FVP tests via echo instead of expect's send
Uses echo as an alternative to expect's send to send stdin to tests. On
the FVP, using expect's send adds a newline to read buffers used in
fscanf calls on stdin which is problematic for testing.

Sending stdin via expect's send adds a newline to the input which breaks
parity between the FVP and other Morello test environments. Using echo
allows us to elide this newline via the -n option.

Change-Id: Ifcc56e6e62cce5e432f87170bfe7e27b91c2430b