Morello: Generate tests bundle and shell script
This patch adds extra target in the test makefile
to generate a portable bundle of unit tests along
with the shell script to run them.

The portability of the dynamically-linked tests is
achieved via proper setting of rpath (this doesn't
change from before) and patchelf-updating the path
to interpreter in the dynamic executables that are
put in the bundle.

The shell script doesn't validate tests output and
is also executing tests sequentially.

Change-Id: I1cd1ea1f0517aecb64fa0e49b2da96f4f2785814