sbrk: Revert morello related changes
sbrk is not a system call but a wrapper around brk. In Musl, it is
explicitly disabled for any argument other than 0. It allows for getting
the current break but not changing it.

In pure-cap brk is disabled, so the upstream version of sbrk will fail
correcly as well. In Aarch64, it will still work as expected for Musl.

Thus the morello-specific changes are not needed anymore now that brk
is properly disabled in the kernel.

Change-Id: Ida9b9341f22874ce113c5fda838c70980fe5394b