malloc: Use alignment builtin for pure-cap and Aarch64
The behavior of enframe between Aarch64 and purecap is inconsistent.
In purecap, it uses a compiler builtin to align the pointer whereas in
Aarch64 it uses a formula.

While testing the preadv and pwritev syscalls with LTP, we observed a
failure due to a misaligned pointer, but only in Aarch64.

It turns out this formula is wrong and can produce unaligned pointers.
For example, 0xfffff7ffc010 align at 0x200 returns 0xFFFF0000080041E0,
which is not aligned at 0x200.

Remove the formula and use the builtin for both pure-cap and Aarch64.

Change-Id: I6ca40c6143c12b4089bccaa14cc420700e6d9f99