Morello: Enable R_MORELLO_CAPINIT relocation
R_MORELLO_CAPINIT relocation is required to generate
capabilities at offsets with a specified size. This needs to be
enabled in the musl dynamic linker to ensure these libraries can
properly be loaded and linked to. R_MORELLO_CAPINIT is
not implemented properly in the musl dynamic linker but this patch aims
to provide a workaround to allow those relocations to be carried out.

Change-Id: Ia8569dc4136bfc057b45d7c8718d29078b2fbb76