Morello: Fix SIGEV_THREAD for timer
Musl would use pointer arithmetic to determine whether timer_t is a
kernel timer or a pthread structure. Pointer arithmetic can generate
invalid capabilities which cannot be reverted back and thus a new type
for timer_t had to be created. The new type is simply a struct with a
boolean attribute to determine if the pointer is to a kernel timer or a
pthread structure.

Change-Id: I47850d40161328df11cd21825f611863ffa125ca