Morello: Moved cmpt blrs mechanism into libc
The trampoline code which performs the cmpt
transition is now copied into an offset in an mmap'd executable buffer
so that any surrounding data is ill-formed code/0.

Labels are used to dynamically calculate the size of the trampoline
code.

Removed the need for an intermediate call to the transition code
by moving struct members into registers manually.

Change-Id: I9489c2c22fe71673ef40bbf99d5f6adac6204385