Morello: Implemented cmpt in xor cipher use case
Used previously created API for compartment mechanism for an xor cipher,
where the user is prompted to enter a key which is stored in a sealed
compartment. The key is then accessed in a function which does the xor
cipher on a given buffer.

Changes have been made to the cmpt API:
- Type is now cmpt_t instead of cap_pair_t
- Separated the asm transition code into call_to_cmpt and trampoline,
  the latter of which is copied into cmpt_t's code page
- Marginally reduced the size of these asm functions
- The allocation of data and code pages is done in one allocation and
  permissions/protections to these pages are set accordingly
- Implemented deallocation of cmpt_t
- Implemented a function which checks the validity of a given cmpt_t

Change-Id: I75052cf4b02c64b651f64f7d693160463d7b8403