Morello: Simplify ELF type definition
When there is a pointer in the Auxiliary vector, each entry needs to
grow to accommodate for the 128 bit capability. The current code was
trying to convince the compiler about the layout of the auxv type
needlessly:
- We do not need to add padding explicitly, since we don't use that in
our code. The compiler will do this automatically, if needed.
- Having an extra entry in the union, of type "pointer", does not hurt
for non-Morello builds, since it's equivalent to the uint64_t entry,
both in size and representation. Having the pointer also for non-Morello
builds actually improves type safety.

So get rid of the extra handling, and add the a_ptr member in both
cases. When a pointer is 64 bits, the union won't be bigger or require
larger alignment, so the struct will look identical. For a Morello
build, the 128-bit capability alignment requirement will force the
compiler to add the padding automatically.

This also allows to get rid of the special handling on the user side,
since now a_ptr will always point the pointer representation of the
entry, which is equivalent to the a_val entry for non-Morello.

Change-Id: Id9fa70509b610ee94ee74b5c9ebfb233b9d419f7