Morello: Re-introduce pointer casts in dlstart.c
In commit 9bdcde58d275 ("Use aux AT_CHERI_* caps in ldso") we dropped
some void* casts, which apparently is fine for clang, but upsets GCC.
Consequently this commit broke compilation with GCC, even for a native
aarch64 build.

This brings back those pointer casts. Doing pointer arithmetics on char
pointers, then assigning the result to a typed pointer is somewhat
fragile, as it could break alignment. It looks like we are fine here,
since the base value is surely properly aligned, and we probably rely on
the aux header values to be correct, so we can "sign off" on this
assignment by adding the cast.

Change-Id: I841c4a85a9bcee3ca724f1608dca4a3f868c8669