ldso: use correct types in dl_iterate_phdr
The default dl_iterate_phdr() implementation was somewhat confused about
the types: AT_PHENT and AT_PHNUM are definitely data types, not
pointers. "base" is used for both, so must be of the bigger type, which
is uintptr_t in Morello. Also GCC complains about void pointer
arithmetic. Last but not least there was a typo in the final assignment
to dlpi_phnum.

Fix all of these mostly by switching to the correct types. Avoiding the
void pointer arithmetic is a bit ugly, but since it is dangerous, we at
least acknowledge this fact with an explicit cast and a comment.

Change-Id: I30abbe32313ce2fa918d75a5607b4a18a5db7fe2