Morello: Fix overlapping mmap allocation for shared libraries
Sections containing BSS requires more virtual memory than filesize on
disk, this is to store uninitialized static variables.

This requires an mmap overlap to MAP_ANONYMOUS the BSS section of
memory. As the capability was casted to a different type, the metadata
was cleared. Under PCuABI, two unreleated capabilities cannot have
overlapping reservations. This patch uses the correct types to maintain
capability metadata and allow the overlap mmap call.

Change-Id: Ic1d9b1f625bc64eb5ae98ee18567fcd96fa5e443