morello: configure: fix purecap flags
the arch flag got changed to -march=morello, but old compilers
and multilib gcc only supports -march=morello+c64 for purecap
so use that, this should continue to work on new compilers,
however recent llvm issues a deprecation warning which we have
to suppress.

and properly test linker flags (they are lld only).