morello: fix dynamic link debug ABI
The dynamic link debug ABI got broken by commits

  commit 78408d48f5b6df450020ea9bcf9d04d3e803e3bd
  Author:     Sherwin da Cruz <sherwin.dacruz@arm.com>
  CommitDate: 2022-09-02 15:56:03 +0000

    Morello: Only set exec/interp RW cap when about to write values

  commit d4c1b3f9f4b372595c15f10b12a051c26581d0a4
  Author:     Harry Ramsey <harry.ramsey@arm.com>
  CommitDate: 2022-10-11 12:29:28 +0100

    Morello: Port dynamic link functions to morello

The link map layout in link.h must match the ldso internal dso struct
and it must not change as it is ABI between ld.so and the debugger.
So remove the new field that got added there and move it elsewhere in
struct dso.

Also we do not want this field in the !purecap case.