Restrict bounds for TLS symbols
In this patch the capability bounds for the TLS symbol
looked up via the dlsym function are set according to
the symbols size information. This is useful to prevent
access to other TLS symbols via the same capability.

This patch also changes how the __tls_get_addr function
works for purecap target. It will accept an array of 3
elements. The 3rd element is purecap-specific and should
contain the size of the object pointed to by the resulting
pointer return from this function.

The implementation of the dl_iterate_phdr (both static and
dynamic) is changed accordingly.

This patch also adds relevant tests.