Morello: Fix dlopen bus error for new TLS
dlopen() bus errors trying to set up the TLS. Pointer for the region was
erroneously mis-calculated using size_t and not realizing the pointer
is not a byte aligned char * etc but is a unitptr_t *. This fixes
that. It also comments the logic and breaks it up for readability.

Signed-off-by: Carsten Haitzler <carsten.haitlzer@arm.com>