Morello: Fix signal test in dynamic case
In purecap, the signal test relies on the address of jump_to_c being
within the bounds of PCC at the point where the signal occurs. As
things stand, this assumption is invalid when running the test as a
dynamic executable, because the signal occurs while executing the
kill() function. Since it lives in libc, there is no guarantee that
jump_to_c (in the main executable) is within the bounds of the
interrupted PCC.

Fix this issue by triggering the signal directly in the main
executable. An easy way to do this is to cause a segfault, for
instance by dereferencing a null pointer.

This problem went unnoticed as the AT_CHERI_*_CAP auxv entries,
which are used to process relocations, remained unbounded in the
Morello Linux kernel. Tightening their bounds, especially that of
AT_CHERI_INTERP_RX_CAP, revealed the invalid assumption made in this
test.