[Morello] compiler-rt build: cmake separating multiple arguments
To ensure argument escape is handled consistently,
pass string arguments containing spaces
via environment variables
instead of via cmake command line.

In the process of the rebase to LLVM-15,
https://git.morello-project.org/morello/llvm-project/-/merge_requests/293
llvm-project has encountered failures which seem to be related to the
CI environment which can't be reproduced easily.
One was worked around by reverting an upstream commit: 891ca8f0
but the remaining ones seem to be more difficult to diagnose.
The musl job seems to be passing separate arguments as one string.

If this does not fix the problem, at least it may rule it out as a
reason for the failure.

Change-Id: I7575c4f840f03929e6d18aed48c87baf78034755