[Morello] build musl-libc with either clang-14 or clang-15
In preparation for building with clang-15
 https://git.morello-project.org/morello/llvm-project/-/merge_requests/293

In configure, pass explicit target to linker, when detecting builtin libraries.

When configuring compiler-rt build:

Use cmake include_directories() which has consistent behaviour,
instead of environment variables CMAKE_<lang>_FLAGS
which can cause problems when containing spaces.

For clang-15, we need to pass -DTARGET_TRIPLE to cmake.
This is not needed when building with clang-14.

build-musl-aarch64 was built (tests built, not run) locally on branch master.
build-musl-x86 was built locally on branch merge-llvm-15.

Change-Id: Icc500d080d455e8b54a9bfcf1417f81899500f55