[Morello] Only if clang version >= 15, specify linker target
In configure, when cross-compiling, pass explicit target to linker, when detecting builtin libraries.
This is necessary when building with clang-15;

Doing this when building with clang-14 caused a pipeline failure on the dev branch. This commit fixes that.

In preparation for building with clang-15.
 https://git.morello-project.org/morello/llvm-project/-/merge_requests/293

Related to
https://jira.arm.com/browse/LLVM-3019 "merge#6 llvm from cheri/dev to morello"

Repo: https://git.morello-project.org/morello/musl-libc
Change-Id: I2eddb9cb780d656be9e99549d7a90c4bf8f90557