[BAKEWELL] Do not rely on provided value for argc
Both on ARM and on RISCV argc for a newly started program is
passed in the same register that is also used for the system
call return value. This is fragile and known to be broken in
at least some kernel version.

Resort to the value provided in the aux vector if present and
only use the register value as a fallback.