ptrace: Provide a proper prototype for ptrace
By default linux uses a var-args prototype for ptrace for
mostly historical reasons. These days virtually all calls
to ptrace are with 4 arguments.

Trying to use var-args will break with CHERI because of the
var-arg ABI issues. Two unsigned longs will be passed on the
stack in a slot that is only one pointer in size and the callee
will than read two arguments as one when reading pointer sized
items from the var-arg store.

To fix this provide a proper 4-args prototype for ptrace instead
and adjust the implementation accordingly. This has the added
benefit that callers will actually get a compile time warning
and not a runtime failure if they try to pass arguments of the
wrong type.