cheri: update cap access perms for Zcherilevels
When the Zcherilevels extension is supported, reading a capability
requires EL, writing a capability requires SL.

Update the permission masks to not strip EL or SL.

Clarify that these are masks for acperm. The bits in the mask mean that
this permission is not removed if the input capability grants it. We
should not remove CL when we prepare a capability that authorizes
capability load or store. Therefore, CL must be included in the masks.