cheri: Fix TLS re-allocation and alignment
The memory layout for the data is:
- The struct dso
- This is followed by a name (pointed to by ->name and located
  at/after the end of the structure in the VLA array buf.
- This is followed by an array of pointers.

In order to satisfy the alignment of the pointer array at most
sizeof(uintptr_t) - 1 bytes must be added to the allocation.
Together with the terminating NUL byte of the string is is a
total of sizeof(uintptr_t).

When aligning make sure that we don't do poniter arithmetic
and align to sizeof(uintptr_t) and not just to sizeof(size_t).