cheri: dynlink: Adhere to FreeBSD CHERI ABI for lsdo list
The list of dynamic libraries loaded into a binary is conveyed
from the dynamic linker in the debug target to the debugger
by means of structures that are located in the target's memory.

The root of the share library information is pointed to by
the DT_DEBUG tag in the dynamic section. The dynamic linker
patches this entry on program startup. The "pointer" is an
unsigned long and not a capability even on CHERI. This is ok
as it is only read by the debugger that can peek into guest
memory anyway.

This pointer points to a struct r_debug which musl incidentally
call "struct debug". This struct contains several pointers and
turn points to a struct link_map. These structures are potentially
used by the dynamic linker and not just the debugger. Thus
pointers contained in these structures must be capabilities.

This changes the ABI between the dynamic linker and the
debugger. FreeBSD's verion of the CHERI enabled gdb has a
fix for this but the fix used does not match our modifications
for the "struct r_debug". The "r_brk" or "bp" field in this
structure is another "pointer" that is only red by the debugger.
Thus it can be and unsigned long and FreeBSD chose to do this.

In order to match the FreeBSD ABI make this field an unsigned long
on linux as well. This is only the musl part of the fix the gdb
side needs fixing for riscv-linux, too.