Fix dirname and basename prototypes
The situation with dirname and basename is complicated because
there are two different versions of basename: The GNU one and
a subtly different version specified in POSIX.

Depending on the implementation these need different prototypes
because the POSIX implementation is allowed to modify the string
passed to it as an argument while the GNU verion uses a static
buffer for the return string making it non-reentrant.

The musl implementation follows the GNU version but specifies
a prototype for basename() without any arguments. This effectively
makes the function a var-args function which does not work well
with CHERI and the compiler complains.

To fix this change prototypes and implementation of basename()
and dirname() to accept a "const char *" parameter.