Fix struct msghdr definition to match the kernel
The struct msghdr used for UDP packets includes two length
fields that the kernel expects to be a size_t. Make the musl
verions of the structure match this expectation and remove
some compat code for 32-bit architectures. These should just
get the correct behaviour if they assign the length fields
as a whole.