[MUSL] Intialize capabilities using relative relocations.
In the static case cbuildcap relocs can be initialized with a cbld
instruction. However when dynamic linking the base address and bounds
needs to be offset by the loaded base address. This cannot be done with
the standard cbld instruction. Until we have a CRelocate instruction we
need to manually inspect the components of the capability and construct
a new one from the info.