morello: ldso: Support new JUMP_SLOT fragment ABI
This is the ABI recently specified in aaelf64-morello[1]. Previously it
was not specified and ended up being identical to aaelf64's ABI, with
the VA of the PLT header in the GOT PLT entry, which is unhelpful for
implementations that wish to support lazy binding. Now it uses the same
fragment format as R_MORELLO_RELATIVE, but this means the addend is now
non-zero and applies to the fragment for the PLT header, not the actual
target, so we must ignore it here.

Note Morello LLVM currently implements the old, accidental ABI, but will
in future switch to the new one, so this commit prepares for that.

[1] https://github.com/ARM-software/abi-aa/pull/249