Fix main's function pointer prototype
__libc_start_main has a function pointer for main that gets passed
to a stage2 function. The function pointer definition for main
does not match the stage2 prototype, which expects a function
pointer with one integer and two pointers. The current main
function pointer is defined with one integer and three pointers.

This commit fixes the mismatch by aligning both definitions and
ensuring that main takes only two pointers.

Fixes: 4d06a81c ("Morello: Switch to argc/argv/envp/auxv supplied in registers")