cheri: Reduce size of sigset_t to match the kernel version
The userland view of sigset_t in musl on riscv has an arbitrary size
of 1024 bit. All communication with the kernel expects _NSIG_WORDS
times the size of a long. This conversion is done by the musl syscall
wrappers whenever a sigset_t is passed from/to the kernel.

In order to allow for future extensions the non-cheri kernel
reserves the same 1024 bit for sigset_t storage in struct ucontext.
However, for CHERI we've consumed some of this reserved space to
make sure that the offset of the saved registers within struct ucontext
is the same for both CHERI and non-CHERI. This significantly simplifies
things in gdb and libunwind among other things.

In the kernel's definition of struct ucontext this reserved space is
not really related to the sigset_t type. The kernel just expects that
its own definition of sigset_t (or __kernel_sigset_t for that matter)
fits into the reserved space. The reserved space is 1024 bits
(plus once the size of a long because of hidden padding).

But then musl comes with its own definition of struct ucontext and
this definition (and some musl code that accesses struct ucontext)
assumes that the musl-declared size of struct sigset_t exactly
matches the size of the reserved space. If this is not true the
offset of the saved registers will be wrong.

By reducing the size of "sigset_t" by two times the size of a long
in the cheri case this assumption is restored.