[BAKEWELL] Fix error running DT_FINI_ARRAY functions
A problem has been seen when some non-PIE executables terminate.
When the program has termination functions a CHERI fault is
raised. The problem occurs when trying to run the first function
recorded in the DT_FINI_ARRAY.

In the linker all dynamic values are expressed as offsets from a `base`
pointer. In the case of non-PIE executables the code is linked
at address 0 (although image base defaults to 64K, ensuring no code
is loaded below that). This causes the dynamic linker to set `base` to
a capability derived from one passed in though the auxvec at
AT_CHERI_EXEC_RX_CAP which covers all the loaded code but with an
address of 0. Normally the dynamic value added to `base` would bring
it back into the capability bounds.

However, when running through the values in the DT_FINI_ARRAY, the
compiler reorders an addition followed by a decrement on a pointer
derived from `base`. When the pointer is 0 this reordering causes the
pointer to go negative, and the capability become unrepresentable.

Signed-off-by: Stuart Menefy <stuart.menefy@codasip.com>